Invalid/Disconnectd IPs are displayed below IPs connected successfully
Each invalid or disconnected IP throws a pop-up warning window in the former version, which is annoying. In the current version, all IPs will be divided into two parts: connected successfully and invalid/disconnected.

Fix video saving naming error in Windows system

When using default video saving name, Date is generated by the host system (the one that main.html is running on). Date is in YYYY-MM-DD format on Linux while it is in MM/DD/YYYY on Windows. And Windows inteperts '/' as 'create a new folder'.

To avoid this issue, divide Date into Year, Month and Day, then re-format them as YYYY-MM-DD.